Control.Pipe.Guarded

Description

Guarded pipes are a drop-in replacement for pipes providing an additional primitive tryAwait, which returns Nothing when the upstream pipe terminates.

This can be used by stateful pipes to yield additional data or compute a return value just before termination.

Note that there is no guarantee that tryAwait will actually return Nothing at some point. If downstream terminates first, the pipe will just terminate immediately.

That means that tryAwait cannot be reliably used as a mechanism to enforce finalization of scarse resources, although it works well to make sure that resources are finalized as soon as possible.

For example:

 foldP :: Monad m => (a -> b -> a) -> a -> Consumer b m a
 foldP f z = go z
   where go z = tryAwait >>= maybe (return z) (go . f z)

uses tryAwait to return the accumulator just before termination. There is no way to implement foldP with regular pipes, since await causes the pipe to terminate as soon as upstream terminates, without giving it a chance to return a result.

Documentation

data Pipe a b m r

type Producer b m r = Pipe Zero b m r

type Consumer a m r = Pipe a Zero m r

type Pipeline m r = Pipe Zero Zero m r

yield :: b -> Pipe a b m ()

await :: Monad m => Pipe a b m a

tryAwait :: Monad m => Pipe a b m (Maybe a)

pipe :: Monad m => (a -> b) -> Pipe a b m r

idP :: Monad m => Pipe a a m r

newtype Lazy m r a b

Constructors

Lazy

Fields

unLazy :: Pipe a b m r

newtype Strict m r a b

Constructors

Strict

Fields

unStrict :: Pipe a b m r

(<+<) :: Monad m => Pipe b c m r -> Pipe a b m r -> Pipe a c m r

(>+>) :: Monad m => Pipe a b m r -> Pipe b c m r -> Pipe a c m r

(<-<) :: Monad m => Pipe b c m r -> Pipe a b m r -> Pipe a c m r

(>->) :: Monad m => Pipe a b m r -> Pipe b c m r -> Pipe a c m r

discard :: Monad m => Pipe a b m r

runPipe :: Monad m => Pipeline m r -> m r